Extract syscall_reboot() handler
